cleanup duplicate code in tests
