+++ tests/components/advantage_air/test_binary_sensor.py
@@ -1,7 +1,6 @@
 """Test the Advantage Air Binary Sensor Platform."""
 from datetime import timedelta
-
-import pytest
+from unittest.mock import AsyncMock
 
 from homeassistant.config_entries import RELOAD_AFTER_UPDATE_DELAY
 from homeassistant.const import STATE_OFF, STATE_ON
@@ -9,20 +8,15 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers import entity_registry as er
 from homeassistant.util import dt as dt_util
 
-from . import add_mock_config, patch_get
+from . import add_mock_config
 
 from tests.common import async_fire_time_changed
 
 
-@pytest.fixture
-def mock_get():
-    """Fixture to patch the Advantage Air async_get method."""
-    with patch_get() as mock_get:
-        yield mock_get
-
-
 async def test_binary_sensor_async_setup_entry(
-    hass: HomeAssistant, entity_registry: er.EntityRegistry, mock_get
+    hass: HomeAssistant,
+    entity_registry: er.EntityRegistry,
+    mock_get: AsyncMock,
 ) -> None:
     """Test binary sensor setup."""
