+++ tests/components/advantage_air/test_climate.py
@@ -1,5 +1,7 @@
 """Test the Advantage Air Climate Platform."""
 
+from unittest.mock import AsyncMock
+
 from advantage_air import ApiError
 import pytest
 
@@ -25,25 +27,14 @@ from homeassistant.core import HomeAssistant
 from homeassistant.exceptions import HomeAssistantError
 from homeassistant.helpers import entity_registry as er
 
-from . import add_mock_config, patch_get, patch_update
-
-
-@pytest.fixture
-def mock_get():
-    """Fixture to patch the Advantage Air async_get method."""
-    with patch_get() as mock_get:
-        yield mock_get
-
-
-@pytest.fixture
-def mock_update():
-    """Fixture to patch the Advantage Air async_get method."""
-    with patch_update() as mock_get:
-        yield mock_get
+from . import add_mock_config, patch_update
 
 
 async def test_climate_myzone_main(
-    hass: HomeAssistant, entity_registry: er.EntityRegistry, mock_get, mock_update
+    hass: HomeAssistant,
+    entity_registry: er.EntityRegistry,
+    mock_get: AsyncMock,
+    mock_update: AsyncMock,
 ) -> None:
     """Test climate platform main entity."""
 
@@ -134,7 +125,10 @@ async def test_climate_myzone_main(
 
 
 async def test_climate_myzone_zone(
-    hass: HomeAssistant, entity_registry: er.EntityRegistry, mock_get, mock_update
+    hass: HomeAssistant,
+    entity_registry: er.EntityRegistry,
+    mock_get: AsyncMock,
+    mock_update: AsyncMock,
 ) -> None:
     """Test climate platform myzone zone entity."""
 
@@ -184,7 +178,10 @@ async def test_climate_myzone_zone(
 
 
 async def test_climate_myauto_main(
-    hass: HomeAssistant, entity_registry: er.EntityRegistry, mock_get, mock_update
+    hass: HomeAssistant,
+    entity_registry: er.EntityRegistry,
+    mock_get: AsyncMock,
+    mock_update: AsyncMock,
 ) -> None:
     """Test climate platform zone entity."""
 
@@ -215,12 +212,15 @@ async def test_climate_myauto_main(
         mock_update.assert_called_once()
 
 
-async def test_climate_async_failed_update(hass: HomeAssistant, mock_get) -> None:
+async def test_climate_async_failed_update(
+    hass: HomeAssistant,
+    mock_get: AsyncMock,
+    mock_update: AsyncMock,
+) -> None:
     """Test climate change failure."""
 
-    with patch_update(side_effect=ApiError) as mock_update, pytest.raises(
-        HomeAssistantError
-    ):
+    with pytest.raises(HomeAssistantError):
+        mock_update.side_effect = ApiError
         await add_mock_config(hass)
         await hass.services.async_call(
             CLIMATE_DOMAIN,
