+++ tests/components/advantage_air/test_sensor.py
@@ -1,7 +1,6 @@
 """Test the Advantage Air Sensor Platform."""
 from datetime import timedelta
-
-import pytest
+from unittest.mock import AsyncMock
 
 from homeassistant.components.advantage_air.const import DOMAIN as ADVANTAGE_AIR_DOMAIN
 from homeassistant.components.advantage_air.sensor import (
@@ -14,27 +13,16 @@ from homeassistant.core import HomeAssistant
 from homeassistant.helpers import entity_registry as er
 from homeassistant.util import dt as dt_util
 
-from . import add_mock_config, patch_get, patch_update
+from . import add_mock_config
 
 from tests.common import async_fire_time_changed
 
 
-@pytest.fixture
-def mock_get():
-    """Fixture to patch the Advantage Air async_get method."""
-    with patch_get() as mock_get:
-        yield mock_get
-
-
-@pytest.fixture
-def mock_update():
-    """Fixture to patch the Advantage Air async_get method."""
-    with patch_update() as mock_get:
-        yield mock_get
-
-
 async def test_sensor_platform(
-    hass: HomeAssistant, entity_registry: er.EntityRegistry, mock_get, mock_update
+    hass: HomeAssistant,
+    entity_registry: er.EntityRegistry,
+    mock_get: AsyncMock,
+    mock_update: AsyncMock,
 ) -> None:
     """Test sensor platform."""
